app: protect out of touchIDs
From practical observation, out of touchIDs may occur. This conflicts with the maximum number of handled touches in iOS devices. Therefore the only reason that may cause out of touchIDs is that the touchIDs are intrinsically non-concurrent safe. This change adds a read-write lock to the touchIDs array.

Fixes fyne-io/fyne#2407.